refactor(dataset): restructure services into ingestion layer
